// frontend/src/client/types.gen.ts
export type MachineNotePublic = {
    title: string;
    content: string;
    shift: NoteShift;
    impact?: (string | null);
    criticality?: NoteCriticality;
    id: string;
    machine_id: string;
    created_by?: (string | null);
    created_at: string;
};
// ...
export type NoteCriticality = 'low' | 'medium' | 'high';

export type NoteShift = 'morning' | 'afternoon' | 'night';